requirements.txt
# Core dependencies
python-binance==1.0.17
ccxt>=4.1.13
pandas>=2.1.4
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
asyncpg==0.30.0              # 🔥 NEW: Required for signal tracking database
redis>=5.0.0
aiohttp>=3.9.1            # 🛠️ Updated for Python 3.12 compatibility
websockets==11.0.3
python-dotenv==1.0.0
pydantic==1.10.7
ta>=0.10.2,<0.11.0
pytest==7.3.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pyyaml>=6.0.1
